Allow Ruby modules to extend Blueprinter::V2::DSL and use the DSL. They can be included by Blueprints.
